[AUDIO] 64 bit issue: use size_t instead of uint32_t in audio_stream_manager
Some uint32_t are still present in audio_stream_manager, despite the
previous patches which aimed to fix the problem.

This patch fixes 2 variables in StreamIn.cpp and StreamOut.cpp.

Tracked-On: https://jira01.devtools.intel.com/browse/GMINL-4991
Reviewed-on: https://android.intel.com/343043
Tested-by: Esclafer, CatherineX <catherinex.esclafer@intel.com>

+++ audio_stream_manager/src/StreamOut.cpp
@@ -286,7 +286,7 @@ void StreamOut::removeEchoReference(struct echo_reference_itfe *reference)
 
 int StreamOut::getPlaybackDelay(ssize_t frames, struct echo_reference_buffer *buffer)
 {
-    uint32_t kernelFrames;
+    size_t kernelFrames;
     int status;
     status = getFramesAvailable(kernelFrames, buffer->time_stamp);
     if (status != android::OK) {
